backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
sqlalchemy==2.0.30
aiomysql==0.2.0
pymysql==1.1.1
alembic==1.13.1
pydantic==2.7.4
pydantic[email]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
pandas==2.2.2
openpyxl==3.1.2
redis==5.0.6
aioredis==2.0.1
httpx==0.27.0
qrcode==7.4.2
pillow==10.3.0
celery==5.4.0
flower==2.0.1
pytest==8.2.2
pytest-asyncio==0.23.7
python-dateutil==2.9.0